+++ src/Traits/ResourceCasting.php
@@ -1,0 +1,237 @@
+<?php
+
+namespace Bfg\Resource\Traits;
+
+use Carbon\CarbonInterface;
+use Illuminate\Contracts\Database\Eloquent\Castable;
+use Illuminate\Contracts\Database\Eloquent\CastsInboundAttributes;
+use Illuminate\Support\Carbon;
+use Illuminate\Support\Collection as BaseCollection;
+use Illuminate\Support\Facades\Date;
+
+trait ResourceCasting
+{
+    /**
+     * @param  string  $name
+     * @param $value
+     * @return mixed
+     */
+    protected function fieldCasting(string $name, $value): mixed
+    {
+        if (isset($this->casts[$name])) {
+            $castType = $this->casts[$name];
+
+            switch ($castType) {
+                case 'int':
+                case 'integer':
+                    return (int) $value;
+                case 'real':
+                case 'float':
+                case 'double':
+                    return $this->fromFloat($value);
+                case 'decimal':
+                    return $this->asDecimal($value, explode(':', $this->casts[$name], 2)[1]);
+                case 'string':
+                    return (string) $value;
+                case 'bool':
+                case 'boolean':
+                    return (bool) $value;
+                case 'object':
+                    return $this->fromJson($value, true);
+                case 'array':
+                case 'json':
+                    return $this->fromJson($value);
+                case 'collection':
+                    return new BaseCollection($this->fromJson($value));
+                case 'date':
+                    return $this->asDate($value);
+                case 'datetime':
+                case 'custom_datetime':
+                    return $this->asDateTime($value);
+                case 'timestamp':
+                    return $this->asTimestamp($value);
+            }
+
+            if (is_string($castType) && class_exists($castType)) {
+                return $this->getClassCastableAttributeValue($name, $value);
+            }
+        }
+
+        return $value;
+    }
+
+    /**
+     * Decode the given JSON back into an array or object.
+     *
+     * @param  string|null  $value
+     * @param  bool  $asObject
+     * @return mixed
+     */
+    public function fromJson(?string $value, bool $asObject = false): mixed
+    {
+        return json_decode($value, ! $asObject);
+    }
+
+    /**
+     * Decode the given float.
+     *
+     * @param  mixed  $value
+     * @return int|float
+     */
+    public function fromFloat(mixed $value): int|float
+    {
+        return match ((string) $value) {
+            'Infinity' => INF,
+            '-Infinity' => -INF,
+            'NaN' => NAN,
+            default => (float) $value,
+        };
+    }
+
+    /**
+     * Return a decimal as string.
+     *
+     * @param  float|string  $value
+     * @param  int  $decimals
+     * @return string
+     */
+    protected function asDecimal(float|string $value, int $decimals): string
+    {
+        return number_format($value, $decimals, '.', '');
+    }
+
+    /**
+     * Return a timestamp as DateTime object with time set to 00:00:00.
+     *
+     * @param  mixed  $value
+     * @return \Illuminate\Support\Carbon
+     */
+    protected function asDate(mixed $value): Carbon
+    {
+        return $this->asDateTime($value)->startOfDay();
+    }
+
+    /**
+     * Return a timestamp as DateTime object.
+     *
+     * @param  mixed  $value
+     * @return bool|Carbon
+     */
+    protected function asDateTime(mixed $value): bool|Carbon
+    {
+        // If this value is already a Carbon instance, we shall just return it as is.
+        // This prevents us having to re-instantiate a Carbon instance when we know
+        // it already is one, which wouldn't be fulfilled by the DateTime check.
+        if ($value instanceof CarbonInterface) {
+            return Date::instance($value);
+        }
+
+        // If the value is already a DateTime instance, we will just skip the rest of
+        // these checks since they will be a waste of time, and hinder performance
+        // when checking the field. We will just return the DateTime right away.
+        if ($value instanceof \DateTimeInterface) {
+            return Date::parse(
+                $value->format('Y-m-d H:i:s.u'), $value->getTimezone()
+            );
+        }
+
+        // If this value is an integer, we will assume it is a UNIX timestamp's value
+        // and format a Carbon object from this timestamp. This allows flexibility
+        // when defining your date fields as they might be UNIX timestamps here.
+        if (is_numeric($value)) {
+            return Date::createFromTimestamp($value);
+        }
+
+        $format = $this->getDateFormat();
+
+        // Finally, we will just assume this date is in the format used by default on
+        // the database connection and use that format to create the Carbon object
+        // that is returned back out to the developers after we convert it here.
+        try {
+            $date = Date::createFromFormat($format, $value);
+        } catch (\InvalidArgumentException $e) {
+            $date = false;
+        }
+
+        return $date ?: Date::parse($value);
+    }
+
+    /**
+     * Get the format for database stored dates.
+     *
+     * @return string
+     */
+    public function getDateFormat(): string
+    {
+        return $this->dateFormat;
+    }
+
+    /**
+     * Return a timestamp as unix timestamp.
+     *
+     * @param  mixed  $value
+     * @return int
+     */
+    protected function asTimestamp(mixed $value): int
+    {
+        return $this->asDateTime($value)->getTimestamp();
+    }
+
+    /**
+     * Cast the given attribute using a custom cast class.
+     *
+     * @param  string  $key
+     * @param  mixed  $value
+     * @return mixed
+     */
+    protected function getClassCastableAttributeValue(string $key, mixed $value): mixed
+    {
+        if (isset($this->classCastCache[$key])) {
+            return $this->classCastCache[$key];
+        } else {
+            $caster = $this->resolveCasterClass($key);
+
+            $value = $caster instanceof CastsInboundAttributes
+                ? $value
+                : $caster->get($this, $key, $value, $this->fields);
+
+            if ($caster instanceof CastsInboundAttributes || ! is_object($value)) {
+                unset($this->classCastCache[$key]);
+            } else {
+                $this->classCastCache[$key] = $value;
+            }
+
+            return $value;
+        }
+    }
+
+    /**
+     * Resolve the custom caster class for a given key.
+     *
+     * @param  string  $key
+     * @return mixed
+     */
+    protected function resolveCasterClass(string $key): mixed
+    {
+        $castType = $this->casts[$key];
+
+        $arguments = [];
+
+        if (is_string($castType) && str_contains($castType, ':')) {
+            $segments = explode(':', $castType, 2);
+
+            $castType = $segments[0];
+            $arguments = explode(',', $segments[1]);
+        }
+
+        if (is_subclass_of($castType, Castable::class)) {
+            $castType = $castType::castUsing($arguments);
+        }
+
+        if (is_object($castType)) {
+            return $castType;
+        }
+
+        return new $castType(...$arguments);
+    }
+}